+++ src/main/java/com/saebyeok/saebyeok/controller/common/CommonController.java
@@ -1,0 +1,22 @@
+package com.saebyeok.saebyeok.controller.common;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.env.Environment;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Arrays;
+
+@RequiredArgsConstructor
+@RestController
+public class CommonController {
+    private final Environment env;
+
+    @GetMapping("/profile")
+    public String getProfile() {
+        return Arrays.stream(env.getActiveProfiles())
+                .filter(it -> it.startsWith("set"))
+                .findFirst()
+                .orElse("set1");
+    }
+}
